enable relay/ir/query_layout
