+++ python/tvm/tir/op.py
@@ -849,26 +849,6 @@ def abs(x, span=None):
     """
     return _ffi_api.abs(x, span)  # type: ignore
 
-
-
-def AutoQuery(N,IC,KH,KW,OC,SH,SW,PH_L,PH_R,PW_L,PW_R,OH,OW):
-    """Get absolute value of the input element-wise.
-
-    Parameters
-    ----------
-    x : PrimExpr
-        Input argument.
-
-    span : Optional[Span]
-        The location of this operator in the source code.
-
-    Returns
-    -------
-    y : PrimExpr
-        The result.
-    """
-    return _ffi_api.AutoQuery(N,IC,KH,KW,OC,SH,SW,PH_L,PH_R,PW_L,PW_R,OH,OW)  # type: ignore
-
     
 def round(x, span=None):
     """Round elements of the array to the nearest integer.
