+++ tutorials/my_trials/test_latency_throughput.sh
@@ -7,6 +7,7 @@ KMP_AFFINITY=granularity=fine,noduplicates,compact,1,0 \
 numactl --physcpubind=0-27 --membind=0 \
 python $HOME/tvm/tutorials/my_trials/benchmark_byoc_dnnl.py 
 
+
 ### check acc
 # numactl --physcpubind=0-27 --membind=0 \
 # python $HOME/tvm/tutorials/my_trials/benchmark_byoc_dnnl.py \
